requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
imbalanced-learn>=0.11.0
optuna>=3.0.0
loguru>=0.7.0
gradio>=4.0.0
pyyaml>=6.0.0
shap>=0.40.0
lime>=0.2.0.0
pymoo>=0.6.0
networkx>=3.0
pyvis>=0.3.1
dowhy>=0.9.0
causalgraphicalmodels>=0.0.4
fairlearn>=0.9.0
